### 2023-01-06 #5046 @editorialbot
[PRE REVIEW]: RxInfer: A Julia package for reactive real-time Bayesian inference
**Submitting author:** @bvdmitri (<a href="http://orcid.org/0000-0001-9655-7986">Dmitry Bagaev</a>)
**Repository:** https://github.com/biaslab/RxInfer.jl
**Branch with paper.md** (empty if default branch): dev-joss-paper
**Version:** v2.4.1
