+++ paper/paper.bib
@@ -266,7 +266,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
     pages = {235-252},
     year = {2022},
     issn = {0888-613X},
-    doi = {https://doi.org/10.1016/j.ijar.2022.06.006},
+    doi = {10.1016/j.ijar.2022.06.006},
     url = {https://www.sciencedirect.com/science/article/pii/S0888613X22000950},
     author = {Semih Akbayrak and İsmail Şenöz and Alp Sarı and Bert {de Vries}},
     keywords = {Factor graphs, Message passing, Natural gradient descent, Probabilistic programming, Variational inference},
