### 2023-01-24 b310e3a
Modify inference for predictions functionality
