+++ src/model.jl
@@ -182,7 +182,6 @@ function ReactiveMP.activate!(model::FactorGraphModel)
     foreach(r -> activate!(r, options), getrandom(model))
     foreach(d -> activate!(d, options), getdata(model))
     foreach(n -> activate!(n, options), getnodes(model))
-
 end
 
 ## constraints 
