### 2023-01-30 b67474c
WIP: update inference function
