### 2023-01-30 1c9eb9b
WIP: Change inference
