+++ Project.toml
@@ -26,7 +26,7 @@ GraphPPL = "3.1.0"
 MacroTools = "0.5.6"
 Optim = "1.0.0"
 ProgressMeter = "1.0.0"
-ReactiveMP = "~3.3.0"
+ReactiveMP = "~3.3.0, 3"
 Reexport = "1.2.0"
 Rocket = "1.6.0"
 TupleTools = "1.2.0"
