+++ Project.toml
@@ -1,7 +1,7 @@
 name = "RxInfer"
 uuid = "86711068-29c9-4ff7-b620-ae75d7495b3d"
 authors = ["Bagaev Dmitry <d.v.bagaev@tue.nl> and contributors"]
-version = "2.5.0"
+version = "2.6.0"
 
 [deps]
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
@@ -26,7 +26,7 @@ GraphPPL = "3.1.0"
 MacroTools = "0.5.6"
 Optim = "1.0.0"
 ProgressMeter = "1.0.0"
-ReactiveMP = "~3.3.0"
+ReactiveMP = "~3.4.0"
 Reexport = "1.2.0"
 Rocket = "1.6.0"
 TupleTools = "1.2.0"
