+++ src/inference.jl
@@ -7,7 +7,7 @@ import DataStructures: CircularBuffer
 
 using MacroTools # for `@autoupdates`
 
-import ReactiveMP: israndom, isdata, isconst, isproxy, isanonymous
+import ReactiveMP: israndom, isdata, isconst, isproxy, isanonymous, allows_missings
 import ReactiveMP: CountingReal
 
 import ProgressMeter
@@ -251,6 +251,7 @@ unwrap_free_energy_option(option::Type{T}) where {T <: Real} = (true, T, Countin
         meta                    = nothing,
         options                 = nothing,
         returnvars              = nothing, 
+        predictvars             = nothing, 
         iterations              = nothing,
         free_energy             = false,
         free_energy_diagnostics = BetheFreeEnergyDefaultChecks,
@@ -274,6 +275,7 @@ For more information about some of the arguments, please check below.
 - `meta  = nothing`: meta specification object, optional, may be required for some models, see `@meta`
 - `options = nothing`: model creation options, optional, see `ModelInferenceOptions`
 - `returnvars = nothing`: return structure info, optional, defaults to return everything at each iteration, see below for more information
+- `predictvars = nothing`: return structure info, optional, see below for more information
 - `iterations = nothing`: number of iterations, optional, defaults to `nothing`, the inference engine does not distinguish between variational message passing or Loopy belief propagation or expectation propagation iterations, see below for more information
 - `free_energy = false`: compute the Bethe free energy, optional, defaults to false. Can be passed a floating point type, e.g. `Float64`, for better efficiency, but disables automatic differentiation packages, such as ForwardDiff.jl
 - `free_energy_diagnostics = BetheFreeEnergyDefaultChecks`: free energy diagnostic checks, optional, by default checks for possible `NaN`s and `Inf`s. `nothing` disables all checks.
@@ -385,6 +387,27 @@ result = inference(
 )
 ```
 
+- ### `predictvars`
+
+`predictvars` specifies the variables which should be predicted. In the model definition these variables are specified
+as datavars, although they should not be passed inside data argument.
+
+Similar to `returnvars`, `predictvars` accepts a `NamedTuple` or `Dict`. There are two specifications:
+- `KeepLast`: saves the last update for a variable, ignoring any intermediate results during iterations
+- `KeepEach`: saves all updates for a variable for all iterations
+
+Example: 
+
+```julia
+result = inference(
+    ...,
+    predictvars = (
+        o = KeepLast(),
+        τ = KeepEach()
+    )
+)
+```
+
 - ### `iterations`
 
 Specifies the number of variational (or loopy belief propagation) iterations. By default set to `nothing`, which is equivalent of doing 1 iteration. 
@@ -481,7 +504,7 @@ function inference(;
     warn = true
 )
     if isnothing(data) && isnothing(predictvars)
-        error("""One of keyword arguments `data` or predictvars must be specified""")
+        error("""One of keyword arguments `data` or `predictvars`` must be specified""")
     end
     __inference_check_dicttype(:initmarginals, initmarginals)
     __inference_check_dicttype(:initmessages, initmessages)
@@ -529,29 +552,36 @@ function inference(;
         returnvars   = Dict(variable => returnoption for (variable, value) in pairs(vardict) if (israndom(value) && !isanonymous(value)))
     end
 
-    # Assuming that the prediction variables are specified as datavars inside @model, e.g. pred = datavar(Float64, n)
-    # Check if `predictvars` is nothing but `data` has missing values
-    if predictvars === nothing
-        predictvars = Dict(
-            variable => KeepLast() for (variable, value) in pairs(vardict) if (isdata(value) && __inference_check_dataismissing(data[variable]) && !isanonymous(value))
-        )
-    else # iterate through vardict and find corresponding variables in predictvars
[…]
