+++ test/test_inference.jl
@@ -552,4 +552,175 @@ end
     end
 end
 
+@testset "Predictions functionality" begin
+
+    # Given the current implementation of RxInfer ecosystem, this rule needs to be predifined 
+    @rule NormalMeanPrecision(:μ, Marginalisation) (m_out::PointMass, q_τ::PointMass) = begin
+        return missing
+    end
+
+    # test #1 (array with missing + predictvars)
+    data = (y = [1.0, -500.0, missing, 100.0],)
+
+    # A simple model for testing that resembles a simple kalman smoother with
+    # random walk state transition
+    @model function model_1(n)
+        x = randomvar(n + 2)
+        o = datavar(Float64, 2) where {allow_missing = true}
+        y = datavar(Float64, n) where {allow_missing = true}
+
+        x_0 ~ NormalMeanPrecision(0.0, 1.0)
+
+        z = x_0
+        for i in 1:n
+            x[i] ~ NormalMeanPrecision(z, 1.0)
+            y[i] ~ NormalMeanPrecision(x[i], 1.0)
+            z = x[i]
+        end
+        x[n + 1] ~ NormalMeanPrecision(z, 1.0)
+        o[1] ~ NormalMeanPrecision(x[n + 1], 1.0)
+        z = x[n + 1]
+        x[n + 2] ~ NormalMeanPrecision(x[n + 1], 1.0)
+        o[2] ~ NormalMeanPrecision(x[n + 2], 1.0)
+    end
+
+    result = inference(model = model_1(length(data[:y])), iterations = 10, data = data, predictvars = (o = KeepLast(),))
+
+    @test all(typeof.(result.predictions[:o]) .<: NormalDistributionsFamily)
+    @test typeof(result.predictions[:y][3]) <: NormalDistributionsFamily
+
+    # test #2 (array with missing + single entry for predictvars)
+    data = (y = [1.0, -10.0, 0.9, missing, missing],)
+
+    @model function model_2(n)
+        x = randomvar(n + 1)
+        o = datavar(Float64) where {allow_missing = true}
+        y = datavar(Float64, n) where {allow_missing = true}
+
+        z ~ NormalMeanPrecision(0.0, 100.0)
+
+        for i in 1:n
+            x[i] ~ NormalMeanPrecision(z, 1.0)
+            y[i] ~ NormalMeanPrecision(x[i], 1.0)
+        end
+        x[n + 1] ~ NormalMeanPrecision(x[n], 1.0)
+        o ~ NormalMeanPrecision(x[n + 1], 1.0)
+    end
+
+    result = inference(model = model_2(length(data[:y])), iterations = 10, data = data, predictvars = (o = KeepEach(),))
+
+    # note we used KeepEach for variable o with BP algorithm (10 iterations), we expect all predicted variables to be equal 
+    @test all(y -> y == result.predictions[:o][1], result.predictions[:o])
+    @test all(typeof.(result.predictions[:y]) .<: NormalDistributionsFamily)
+
+    # test #3 (array + single entry for predictvars)
+    data = (y = [1.0, -10.0, 0.9],)
+    @model function model_3(n)
+        x = randomvar(n + 1)
+        o = datavar(Float64) where {allow_missing = true}
+        y = datavar(Float64, n) where {allow_missing = true}
+
+        z ~ NormalMeanPrecision(0.0, 100.0)
+
+        for i in 1:n
+            x[i] ~ NormalMeanPrecision(z, 1.0)
+            y[i] ~ NormalMeanPrecision(x[i], 1.0)
+        end
+        x[n + 1] ~ NormalMeanPrecision(x[n], 1.0)
+        o ~ NormalMeanPrecision(x[n + 1], 1.0)
[…]
