+++ examples/Project.toml
@@ -12,7 +12,9 @@ Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
+ReactiveMP = "a194aa59-28ba-4574-a09c-4a745416d6e3"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
+Rocket = "df971d30-c9d6-4b37-b8ff-e965b2cb3a40"
 RxInfer = "86711068-29c9-4ff7-b620-ae75d7495b3d"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
