+++ paper/paper.bib
@@ -160,6 +160,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 @article{podusenko_aida_2021,
 	title = {{AIDA}: {An} {Active} {Inference}-based {Design} {Agent} for {Audio} {Processing} {Algorithms}},
 	shorttitle = {{AIDA}},
+	doi = {10.3389/frsip.2022.842477},
 	url = {http://arxiv.org/abs/2112.13366},
 	abstract = {In this paper we present AIDA, which is an active inference-based agent that iteratively designs a personalized audio processing algorithm through situated interactions with a human client. The target application of AIDA is to propose on-the-spot the most interesting alternative values for the tuning parameters of a hearing aid (HA) algorithm, whenever a HA client is not satisfied with their HA performance. AIDA interprets searching for the "most interesting alternative" as an issue of optimal (acoustic) context-aware Bayesian trial design. In computational terms, AIDA is realized as an active inference-based agent with an Expected Free Energy criterion for trial design. This type of architecture is inspired by neuro-economic models on efficient (Bayesian) trial design in brains and implies that AIDA comprises generative probabilistic models for acoustic signals and user responses. We propose a novel generative model for acoustic signals as a sum of time-varying auto-regressive filters and a user response model based on a Gaussian Process Classifier. The full AIDA agent has been implemented in a factor graph for the generative model and all tasks (parameter learning, acoustic context classification, trial design, etc.) are realized by variational message passing on the factor graph. All verification and validation experiments and demonstrations are freely accessible at our GitHub repository.},
 	urldate = {2022-01-04},
@@ -195,6 +196,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 	abstract = {The switching autoregressive model is a flexible model for signals generated by non-stationary processes. Unfortunately, evaluation of the exact posterior distributions of the latent variables for a switching autoregressive model is analytically intractable, and this limits the applicability of switching autoregressive models in practical signal processing tasks. In this paper we present a message passing-based approach for computing approximate posterior distributions in the switching autoregressive model. Our solution tracks approximate posterior distributions in a modular way and easily extends to more complicated model variations. The proposed message passing algorithm is verified and validated on synthetic and acoustic data sets respectively.},
 	booktitle = {2022 30th {European} {Signal} {Processing} {Conference} ({EUSIPCO})},
 	author = {Podusenko, Albert and van Erp, Bart and Bagaev, Dmitry and şenöz, Ïsmail and de Vries, Bert},
+	doi = {10.23919/EUSIPCO55093.2022.9909828},
 	month = aug,
 	year = {2022},
 	note = {ISSN: 2076-1465},
@@ -209,6 +211,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 	abstract = {This paper bridges the gap in the literature between neural networks and probabilistic graphical models. Invertible neural networks are incorporated in factor graphs and inference in this model is described by linearization of the network. Consequently, hybrid probabilistic inference in the model is realized through message passing with local constraints on the Bethe free energy. We provide the local Bethe free energy for the invertible neural network node, which allows for evaluation of the performance of the entire probabilistic model. Experimental results show effective hybrid inference in a neural network-based probabilistic model for a binary classification task, paving the way towards a novel class of machine learning models.},
[…]
