### 2023-02-17 8b53890
Added make_node for varargs{datavars}, with tests for operations on datavars. Succesful construction of proxy datavars, failure during inference.
