+++ test/runtests.jl
@@ -230,6 +230,8 @@ end
     addtests(testrunner, "models/autoregressive/test_ar.jl")
     addtests(testrunner, "models/autoregressive/test_lar.jl")
 
+    addtests(testrunner, "models/datavars/test_fn_datavars.jl")
+
     addtests(testrunner, "models/mixtures/test_gmm_univariate.jl")
     addtests(testrunner, "models/mixtures/test_gmm_multivariate.jl")
     addtests(testrunner, "models/mixtures/test_switch.jl")
