### 2023-02-18 2576ba7
Added getpointmass method on type Float64 for mapped datavar constructor.
