### 2023-02-20 biaslab/RxInfer.jl#62 @alstat
Provide metrics for claimed accuracy vs Turing.jl
Hi, this is in connection with [JOSS#5161](https://github.com/openjournals/joss-reviews/issues/5161). I've reviewed the benchmark notebook provided. I can confirm that the RxInfer.jl is much faster than Turing.jl, but I'm not sure with the accuracy. While the plots in the README seems to suggest RxInfer.jl to be more accurate, subjectively, when I ran the notebook I found that it is not set to a specific seed, so I got a different plots (see below), and it seems RxInfer.jl and Turing.jl are both close in accuracy especially for Turing.jl's HMC(0.05, 10.0, 1000), subjectively. So, I suggest that you provide metric for measuring the accuracy for this claim, maybe RMSE or MAE would do. Also, is this something theoretically true? I'm just thinking if this may only be true for this example but for others  both RxInfer.jl and Turing.jl maybe in the same range of accuracy after all, unless there is something in the theory that would suggest otherwise or in Turing.jl's implementation. In cases where it is theoretically supported, then maybe we emphasize briefly in the README or the doc on why it is more accurate (haven't read other parts of the documentation yet in case this is already emphasized), otherwise, we may caution that this is only true for the said example and must be tested for more examples or yet to be confirmed theoretically.
