### 2023-02-20 biaslab/RxInfer.jl#63 @alstat
Brief emphasis on the use of aliases for Distributions
Hi again, I followed some of the examples in the documentation, and initially thought that RxInfer.jl has its own implementation of probability distributions, for example for `MVGaussian`, `NormalMeanVariance `, etc. But upon inspecting the [source code](https://github.com/biaslab/RxInfer.jl/blob/ffd6787c2d8eb8df929d0f6c1748d280bdb7b2fc/src/graphppl.jl#L464), I found out that these are just aliases to the probability distributions defined in Distributions.jl (please confirm). If this is the case, maybe we can emphasize in the doc that RxInfer.jl uses Distributions.jl but with extended aliases. The idea here is to avoid misleading impression of learning new APIs for probability distributions from new users who are already familiar with the Distributions.jl.

[JOSS#5161](https://github.com/openjournals/joss-reviews/issues/5161)
