### 2023-02-20 biaslab/RxInfer.jl#63 @bartvanerp
Hi @alstat!

Thanks for the feedback! Your observation is partly correct. Indeed for certain distributions we rely in `ReactiveMP.jl` on the implementation from `Distributions.jl`. These cases include for example the `Beta` and `Wishart` distributions.
*However, specifically for the Gaussian/Normal case we have custom implementations that yield a higher computational efficiency and improved stability. Our aliases for these distributions therefore do not correspond to the implementations from `Distributions.jl`.* This is mainly because of the following 3 reasons:
1. `Distributions.jl` constructs normal distributions by saving the corresponding covariance matrices in a `PDMat` object from `PDMats.jl`. This construction always computes the Cholesky decompositions of the covariance matrices, which is very convenient for sampling-based procedures. However, in `RxInfer.jl` we mostly base our computations on analytical expressions which do not always need to compute the Cholesky decomposition. In order to reduce the overhead that `Distributions.jl` introduces, we therefore have custom implementations.
2. Depending on the update rules, we might favor different parameterizations of the normal distributions. `ReactiveMP.jl` has quite a variety in parameterizations that allow us to efficient computations where we convert between parameterizations as little as possible.
3. In certain situations we value stability a lot, especially when inverting matrices. `PDMats.jl`, and hence `Distributions.jl`, is not capable to fulfill all needs that we have here. Therefore we use `PositiveFactorizations.jl` to cope with the corner-cases.

Because of this partial dependency, we have added functionality in line with what `Distributions.jl` implements to make it convenient for `Distributions.jl` users. I agree that this might not be as well represented in our documentation though. For this reason I will update the documentation accordingly to reflect your feedback. I will leave this issue open as a reminder and close it once the documentation is updated.
