### 2023-02-20 biaslab/RxInfer.jl#63 @bvdmitri
@bartvanerp I think your explanation is very descriptive, we should indeed put that in our documentation
