### 2023-02-20 biaslab/RxInfer.jl#62 @bvdmitri
Another reference is https://www.softwareimpacts.com/article/S2665-9638(22)00042-2/pdf
and the repository with experiments https://github.com/biaslab/ReactiveMPPaperExperiments (that includes the model from the README plus two more)
