### 2023-02-20 f04ce02
map on combined datavars now outputs Message of PointMass type. Inference successful. Added tests.
