+++ test/models/datavars/test_fn_datavars.jl
@@ -18,17 +18,28 @@ include(joinpath(@__DIR__, "..", "..", "utiltests.jl"))
     y ~ Normal(mean = x, variance = 1.0)
 end
 
-## Inference definition
-function datavars_inference(modelfn, adata, bdata, ydata)
+@model function ratio_datavars_as_gaussian_mean()
+
+    a = datavar(Float64)
+    b = datavar(Float64)
+    y = datavar(Float64)
+
+    x ~ Normal(mean = a/b, variance = 1.0)
+    y ~ Normal(mean = x, variance = 1.0)
+end
+
+# Inference function
+function fn_datavars_inference(modelfn, adata, bdata, ydata)
     return inference(
         model = modelfn(),
         data = (a = adata, b = bdata, y = ydata),
-        # free_energy = true
+        free_energy = true
     )
 end
 
 @testset "datavars" begin
 
+    # Test model construction
     model,_ = create_model(sum_datavars_as_gaussian_mean())
     isproxy_vars = [ReactiveMP.isproxy(var) for var in model.variables.data]
     @test length(isproxy_vars) == 4
@@ -39,12 +50,23 @@ end
     bdata = 1.0
     ydata = 0.0
 
-    result = datavars_inference(sum_datavars_as_gaussian_mean, adata, bdata, ydata)
+    # Test inference in model with linear operation on datavars
+    result = fn_datavars_inference(sum_datavars_as_gaussian_mean, adata, bdata, ydata)
     xres = result.posteriors[:x]
-    # fres = result.free_energy
+    fres = result.free_energy
 
     @test typeof(xres) <: NormalDistributionsFamily
     @test isapprox(mean(xres), 1.5, atol = 0.1)
+    @test isapprox(fres[end], 3.51551, atol = 0.1)
+
+    # Test inference in model with linear operation on datavars
+    result = fn_datavars_inference(ratio_datavars_as_gaussian_mean, adata, bdata, ydata)
+    xres = result.posteriors[:x]
+    fres = result.free_energy
+
+    @test typeof(xres) <: NormalDistributionsFamily
+    @test isapprox(mean(xres), 1.0, atol = 0.1)
+    @test isapprox(fres[end], 2.26551, atol = 0.1)
 
 end
 
