+++ src/model.jl
@@ -393,18 +393,15 @@ function ReactiveMP.make_node(model::FactorGraphModel, options::FactorNodeCreati
     end
 end
 
-import ReactiveMP: getpointmass
-ReactiveMP.getpointmass(point::Float64) = point
-
 function ReactiveMP.make_node(model::FactorGraphModel, options::FactorNodeCreationOptions, fform, autovar::AutoVar, args::Vararg{<:ReactiveMP.DataVariable})
     if isstochastic(sdtype(fform))
         var  = make_autovar(model, ReactiveMP.EmptyRandomVariableCreationOptions, ReactiveMP.name(autovar), true)
         node = ReactiveMP.make_node(model, options, fform, var, args...) # add! is inside
         return node, var
     else
-        stream = combineLatest(ReactiveMP._getmarginal.(args)..., strategy=PushNew()) |> map(Any, (d) -> fform(ReactiveMP.getpointmass.(ReactiveMP.getdata(d))...))
-        options = DataVariableCreationOptions(stream)
-        var = push!(model, ReactiveMP.datavar(options, ReactiveMP.name(autovar), Any))
+        combined_datavars = combineLatest(ReactiveMP.getmarginal.(args, IncludeAll())..., strategy=PushNew())
+        mapped_datavars = combined_datavars |> map(Any, (d) -> Message(PointMass.(fform(ReactiveMP.getpointmass.(ReactiveMP.getdata(d))...)), false, false, nothing))
+        var = push!(model, ReactiveMP.datavar(DataVariableCreationOptions(mapped_datavars), ReactiveMP.name(autovar), Any))
         return nothing, var
     end
 end
