### 2023-02-22 biaslab/RxInfer.jl#62 @alstat
Hi @bvdmitri, thank you for these details and the references! I agree that it is not a straightforward comparison, but as you suggest, we can just improve the wording in the README, maybe one more sentence to the current statement would do, something like:

RxInfer.jl not only beats generic-purpose Bayesian inference methods, executes faster, and scales better, but also provides more accurate results for various complex problems. This is especially true for conjugate models, see [add references here]. Check out our [examples](https://biaslab.github.io/RxInfer.jl/stable/examples/overview/)!

Or if you have other better wording.
