+++ docs/src/contributing/new-release.md
@@ -1,7 +1,7 @@
 # [Publishing a new release](@id contributing-new-release)
 
-Please first read the general [Contributing](@ref contributing-overview) section.
-Also please read the [FAQ](https://github.com/JuliaRegistries/General#faq) section in the Julia General registry.
+Please read first the general [Contributing](@ref contributing-overview) section.
+Also, please read the [FAQ](https://github.com/JuliaRegistries/General#faq) section in the Julia General registry.
 
 ## Start the release process
 
@@ -12,4 +12,4 @@ In order to start the release process a person with the associated permissions s
 
 ![Release comment](../assets/img/release_comment.png)
 
-The Julia Registrator bot should automaticallly register a requst for the new release. After all checks will be passed on the Julia Registrator side the new release will be published and tagged automatically.
+The Julia Registrator bot should automaticallly register a request for the new release. Once all checks have passed on the Julia Registrator's side, the new release will be published and tagged automatically.
