### 2023-02-22 biaslab/RxInfer.jl#65 @alstat
Statement of need in software documentation
The statement of need in the JOSS manuscript is very clear. However, one review criteria is that the software documentation itself should also contain a statement of need. I don't expect it to be as long as that in the manuscript; a sentence or two in the README or manual could explain what need this package fills. [JOSS#5161](https://github.com/openjournals/joss-reviews/issues/5161)
