### 2023-02-22 6064cb8
docs(): improve wording in README & add statement of need
