### 2023-02-22 9cf1151
docs(): fixes, typos and other text improvements
