+++ docs/src/index.md
@@ -44,20 +44,18 @@ Given a probabilistic model, RxInfer allows for an efficient message-passing bas
 
 ## Why RxInfer
 
-Many important AI applications, such as audio-processing, self-driving vehicles, weather forecasting, extended reality video processing, and others require continually solving an inference task in sophisticated probabilistic models with a large number of latent variables. 
-Often, the inference task in these applications must be performed continually and in real time in response to new observations. 
-Popular MC-based inference methods, such as No U-Turn Samples (NUTS) or Hamiltonian Monte Carlo (HMC), rely on computationally heavy sampling procedures that do not scale well to probabilistic models with thousands of latent states. 
-Therefore, MC-based inference is practically not suitable for real-time applications. 
-While the alternative variational inference (VI) method promises to scale better to large models than sampling-based inference, VI requires the derivation of gradients of the "variational Free Energy" cost function. 
-For large models, manual derivation of these gradients might be not feasible, while automated "black-box" gradient methods do not scale either because they are not capable of taking advantage of sparsity or conjugate pairs in the model. 
+Many important AI applications, including audio processing, self-driving vehicles, weather forecasting, and extended-reality video processing require continually solving an inference task in sophisticated probabilistic models with a large number of latent variables.
+Often, the inference task in these applications must be performed continually and in real-time in response to new observations.
+Popular MC-based inference methods, such as the No U-Turn Sampler (NUTS) or Hamiltonian Monte Carlo (HMC) sampling, rely on computationally heavy sampling procedures that do not scale well to probabilistic models with thousands of latent states.
+Therefore, MC-based inference is practically not suitable for real-time applications.
+While the alternative variational inference method (VI) promises to scale better to large models than sampling-based inference, VI requires the derivation of gradients of a "variational Free Energy" cost function.
+For large models, manual derivation of these gradients might not be feasible, while automated "black-box" gradient methods do not scale either because they are not capable of taking advantage of sparsity or conjugate pairs in the model.
 Therefore, while Bayesian inference is known as the optimal data processing framework, in practice, real-time AI applications rely on much simpler, often ad hoc, data processing algorithms.
 
-`RxInfer` provides utility to run efficient Bayesian inference in sophisticated probabilistic models, 
-takes advantages of conjugate relationships in probabilistic models, and focuses to perform real-time Bayesian inference 
-in large state-space models with thousands of latent variables. In addition, `RxInfer` provides a straightforward 
-way to extend its functionality with custom factor nodes and message passing update rules. The engine is capable of running
-various Bayesian inference algorithms in different parts of the factor graph of a single probabilistic model. This makes it easier 
-to explore different "what-if" scenarious and enables very efficient inference in specific cases.
+RxInfer aims to remedy these issues by running efficient Bayesian inference in sophisticated probabilistic models,
+taking advantage of local conjugate relationships in probabilistic models, and focusing on real-time Bayesian inference in large state-space models with thousands of latent variables. In addition, RxInfer provides a straightforward way to extend its functionality with custom factor nodes and message passing update rules. The engine is capable of running
+various Bayesian inference algorithms in different parts of the factor graph of a single probabilistic model. This makes it easier
+to explore different "what-if" scenarios and enables very efficient inference in specific cases.
 
 ## Ecosystem
 
