### 2023-02-22 06f8391
isproxy method on datavar moved to ReactiveMP
