### 2023-02-22 d457a20
Output of make_node map changed to Marginal
