### 2023-02-22 #5161 @bvdmitri
The text has been improved a bit, mostly typos. See https://github.com/biaslab/RxInfer.jl/commit/6e34b5897c2a0bbd7789266698c1865ebb79a7be
