+++ test/models/datavars/test_fn_datavars.jl
@@ -8,66 +8,84 @@ using RxInfer, BenchmarkTools, Random, Plots, Dates, LinearAlgebra, StableRNGs
 include(joinpath(@__DIR__, "..", "..", "utiltests.jl"))
 
 ## Model definition
-@model function sum_datavars_as_gaussian_mean()
-
+@model function sum_datavars_as_gaussian_mean_1()
     a = datavar(Float64)
     b = datavar(Float64)
     y = datavar(Float64)
 
-    x ~ Normal(mean = a+b, variance = 1.0)
+    x ~ Normal(mean = a + b, variance = 1.0)
+    y ~ Normal(mean = x, variance = 1.0)
+end
+
+@model function sum_datavars_as_gaussian_mean_2()
+    a = datavar(Float64)
+    b = datavar(Float64)
+    c = constvar(0.0) # Should not change the result
+    y = datavar(Float64)
+
+    x ~ Normal(mean = (a + b) + c, variance = 1.0)
     y ~ Normal(mean = x, variance = 1.0)
 end
 
 @model function ratio_datavars_as_gaussian_mean()
-
     a = datavar(Float64)
     b = datavar(Float64)
     y = datavar(Float64)
 
-    x ~ Normal(mean = a/b, variance = 1.0)
+    x ~ Normal(mean = a / b, variance = 1.0)
     y ~ Normal(mean = x, variance = 1.0)
 end
 
 # Inference function
 function fn_datavars_inference(modelfn, adata, bdata, ydata)
-    return inference(
-        model = modelfn(),
-        data = (a = adata, b = bdata, y = ydata),
-        free_energy = true
-    )
+    return inference(model = modelfn(), data = (a = adata, b = bdata, y = ydata), free_energy = true)
 end
 
 @testset "datavars" begin
-
-    # Test model construction
-    model,_ = create_model(sum_datavars_as_gaussian_mean())
-    isproxy_vars = [ReactiveMP.isproxy(var) for var in model.variables.data]
-    @test length(isproxy_vars) == 4
-    @test sum(isproxy_vars) == 1
-    @test ReactiveMP.isproxy(model.variables.data[isproxy_vars])
-
     adata = 2.0
     bdata = 1.0
     ydata = 0.0
 
     # Test inference in model with linear operation on datavars
-    result = fn_datavars_inference(sum_datavars_as_gaussian_mean, adata, bdata, ydata)
+    result = fn_datavars_inference(sum_datavars_as_gaussian_mean_1, adata, bdata, ydata)
+    model = result.model
     xres = result.posteriors[:x]
     fres = result.free_energy
 
+    isproxy_vars = [ReactiveMP.isproxy(var) for var in model.variables.data]
+    @test length(isproxy_vars) == 4
+    @test sum(isproxy_vars) == 1
+    @test ReactiveMP.isproxy(model.variables.data[isproxy_vars])
+    @test typeof(xres) <: NormalDistributionsFamily
+    @test isapprox(mean(xres), 1.5, atol = 0.1)
+    @test isapprox(fres[end], 3.51551, atol = 0.1)
+
+    result = fn_datavars_inference(sum_datavars_as_gaussian_mean_2, adata, bdata, ydata)
+    model = result.model
+    xres = result.posteriors[:x]
+    fres = result.free_energy
+
+    isproxy_vars = [ReactiveMP.isproxy(var) for var in model.variables.data]
+    @test length(isproxy_vars) == 5
+    @test sum(isproxy_vars) == 2
+    @test ReactiveMP.isproxy(model.variables.data[isproxy_vars])
     @test typeof(xres) <: NormalDistributionsFamily
     @test isapprox(mean(xres), 1.5, atol = 0.1)
[…]
