### 2023-02-22 b43c429
Added broken test for indexed datavar map.
