### 2023-02-22 e8d292f
Added isused flag to DataVariable to suppress warning that proxy datavar is not connected to other nodes. Added tests.
