### 2023-02-23 biaslab/RxInfer.jl#62 @bvdmitri
See #66
