+++ .github/workflows/CI.yml
@@ -1,13 +1,15 @@
 name: CI
 on:
   pull_request:
-    types: [ready_for_review,closed,reopened]
+    types: [ready_for_review,closed,reopened,synchronize]
   pull_request_review:
-    types: [submitted]
+    types: [submitted,edited]
   push:
     branches:
       - 'main'
     tags: '*'
+  check_run:
+    types: [rerequested]
   schedule:
     - cron: '0 8 * * 1' # run the cron job one time per week on Monday 8:00 AM
 concurrency:
