+++ docs/src/contributing/overview.md
@@ -83,8 +83,8 @@ make test
 
 - `make help`: Shows help snippet
 - `make test`: Run tests, supports extra arguments
-  - `make test testest="distributions:normal_mean_variance"` would run tests only from `distributions/test_normal_mean_variance.jl`
-  - `make test testset="distributions:normal_mean_variance models:lgssm"` would run tests both from `distributions/test_normal_mean_variance.jl` and `models/test_lgssm.jl`
+  - `make test test_args="distributions:normal_mean_variance"` would run tests only from `distributions/test_normal_mean_variance.jl`
+  - `make test test_args="distributions:normal_mean_variance models:lgssm"` would run tests both from `distributions/test_normal_mean_variance.jl` and `models/test_lgssm.jl`
   - `make test dev=true` would run tests while using `dev-ed` versions of core packages
 - `make devtest`: Alias for the `make test dev=true ...`
 - `make docs`: Compile documentation
