### 2023-02-23 df06f80
ci(): do not run CI on closed PRs
