### 2023-02-24 biaslab/RxInfer.jl#62 @alstat
Thank you, @bvdmitri! Looks good to me.
