### 2023-02-24 #5161 @alstat
Hi @bvdmitri, one of the criteria for the paper is to also describe how RxInfer.jl is compared to other software, I only require comparison against Julia frameworks, but you are welcome to compare it to Stan or TensorFlow Probability.
