### 2023-02-24 biaslab/RxInfer.jl#71 @bvdmitri
Add a section in the documentation that compares RxInfer to other well-known Bayesian inference packages
Part of the https://github.com/openjournals/joss-reviews/issues/5161 review process, but we must do that in any case as it improves general presentability of the package and is very convenient for users.
