### 2023-02-24 #5161 @alstat
Copy @bvdmitri, note that the criteria mentioned is for the JOSS paper itself, but good to have if it is also added in the documentation.
