### 2023-03-01 440186b
Fix README formulas
Fixes #81
