### 2023-03-01 8cf280e
Use GitHub latex engine to render equations
