### 2023-03-02 3a0055d
paper(): compare to Turing/Stan
