+++ paper/paper.bib
@@ -311,3 +311,33 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 	editor = {Steve Brooks and Andrew Gelman and Galin Jones and Xiao-Li Meng},
 	title = {Handbook of Markov Chain Monte Carlo}
 }
+
+@inproceedings{ge2018t,
+  author    = {Hong Ge and
+               Kai Xu and
+               Zoubin Ghahramani},
+  title     = {Turing: a language for flexible probabilistic inference},
+  booktitle = {International Conference on Artificial Intelligence and Statistics,
+               {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
+               Spain},
+  pages     = {1682--1690},
+  year      = {2018},
+  url       = {http://proceedings.mlr.press/v84/ge18b.html},
+  biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
+}
+
+@inproceedings{standevteam,
+  author    = {Stan Development Team},
+  title     = {Stan Modeling Language Users Guide and Reference Manual, Version 2.31},
+  year      = {2022},
+  url       = {https://mc-stan.org},
+  biburl    = {https://mc-stan.org},
+}
+
+@inproceedings{standevteam_jl,
+  author    = {Stan.jl Development Team},
+  title     = {Stan Modeling Language in Julia, Version 10.3.2},
+  year      = {2022},
+  url       = {https://github.com/StanJulia/Stan.jl},
+  biburl    = {https://github.com/StanJulia/Stan.jl},
+}
