### 2023-03-02 c60cc05
paper(): fix Stan bib entry
