+++ paper/paper.bib
@@ -327,7 +327,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 }
 
 @inproceedings{standevteam,
-  author    = {Stan Development Team},
+  author    = {{Stan Development Team}},
   title     = {Stan Modeling Language Users Guide and Reference Manual, Version 2.31},
   year      = {2022},
   url       = {https://mc-stan.org},
@@ -335,7 +335,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 }
 
 @inproceedings{standevteam_jl,
-  author    = {Stan.jl Development Team},
+  author    = {{Stan.jl Development Team}},
   title     = {Stan Modeling Language in Julia, Version 10.3.2},
   year      = {2022},
   url       = {https://github.com/StanJulia/Stan.jl},
