### 2023-03-02 36b5a4a
Update inference to allow user interruption
a draft code of allowing inference function failed, especially for user interruption
