### 2023-03-03 83c31ee
add a test demo for allow_failed
