### 2023-03-05 #5161 @alstat
Comment on citation:

Line 109, please change `[@sarkka_bayesian_2013]` to `@sarkka_bayesian_2013`, so that it will be rendered as Särkkä (2013) instead of (Särkkä, 2013).
