### 2023-03-06 2642b0d
paper(): improve citation to Sarkka
