### 2023-03-10 f0ec70d
feat: allow to halt iterations based on callback return value
