### 2023-03-17 #5161 @alstat
Hi @jmschrei, I'm done reviewing the package already. I think the team behind RxInfer.jl exerted quite an effort to come up with a quality software package for Julia, specifically as additional tool for doing Bayesian inference with the advantage of being fast for real-time applications. The core APIs is not only intuitive but also consistent from model, constraint, to meta specification. On top of this, they have provided an extensive examples that readers can explore on the capability of RxInfer.jl. Lastly, the authors of the package have been responsive enough for all my queries, and they have addressed all of the issues I raised. Thank you.
