### 2023-03-19 d6c22d0
update allow failed
Update the inference function and add a new function in actor.jl. This new function is used to remove the #undef values in Free energy caused by user interruptions.
