### 2023-03-20 394cab7
remove testing notebook and change throw()
