### 2023-03-23 0d3af20
test: run benchmarks on a single CPU
