+++ test/models/iid/test_mv_iid_precision_known_mean.jl
@@ -57,7 +57,7 @@ end
         p = contour!(p, X, Y, (x, y) -> pdf(MvNormalMeanPrecision(m, P), [x, y]), label = "Real")
     end
 
-    @test_benchmark "models" "iid_mv_wishart_known_mean" inference_mv_wishart_known_mean($m, $data, $n, $d)
+    @test_benchmark "models" "iid_wishart_known_mean" inference_mv_wishart_known_mean($m, $data, $n, $d)
 end
 
 end
