### 2023-03-24 #5161 @dhvalden
hi @jmschrei, I cannot but agree with @alstat. For what I see from the extensive library of examples, its quite clear that the authors have put considerable effort to make this library a very accessible and robust one. I have run several examples in my machine with zero problems related to the library, all of this in a Linux Fedora machine. I have used Turing.jl in the past and this library looks like a strong competition even with considering the real-time selling point. I have no comments from my limited end user perspective more that congratulating the authors.
