### 2023-03-26 a58d87f
Update new-example.md
Improve wording a little bit
