+++ docs/src/contributing/new-example.md
@@ -6,14 +6,17 @@ In order to add a new example simply create a new Jupyter notebook with your exp
 
 After that it is necessary to modify the `examples/.meta.jl` file. See the comments in this file for more information.
 
-1. Make sure that the very first cell of the notebook contains ONLY `# <title>` in it and has markdown type. This is important for link generation in the documentation
-2. Paths must be local and cannot be located in subfolders
-3. Description is used to pre-generate an Examples page overview in the documentation
-4. Use hidden option to not include a certain example in the documentation (build will still run to ensure the example runs)
-5. Name `Overview` is reserved, please do not use it
-6. Use \$\$\begin{aligned} (note the same line, otherwise formulas will not render correctly in the documentation)
-                   <latex formulas here>
-                   \end{aligned}\$\$ (on the same line (check other examples if you are not sure)
+1. Make sure that the very first cell of the notebook contains ONLY `# <title>` in it and has markdown type. This is important for link generation in the documentation.
+2. The `path` option must be a local path and cannot contain subfolders.
+3. The `description` option is used to pre-generate the `Examples` page overview in the documentation.
+4. Set `hidden = true` option to hide a certain example in the documentation (the example will be executed to ensure it runs without errors).
+5. Please do no use `Overview` as a name for the new example, the `Overview` is reserved.
+6. Use the following template for formulas, not that `$$` and both `\begin` and `\end` commands are on the same line (check other examples if you are not sure). This is important, because otherwise formulas may not render correctly. Inline formulas may use `$...$` template.
+```
+$$\begin{aligned}
+      <latex formulas here>
+\end{aligned}$$
+``` 
 7. Notebooks and plain Julia have different scoping rules for global variables, if it happens so that examples generation fails due to `UndefVarError` and scoping issues use `let ... end` blocks to enforce local scoping (see `Gaussian Mixtures Multivariate.ipynb` as an example)
 
 8. All examples must use and activate local `Project.toml` in the second cell (see `1.`), if you need some package add it to the `(examples)` project
