+++ docs/src/contributing/new-example.md
@@ -8,10 +8,10 @@ After that it is necessary to modify the `examples/.meta.jl` file. See the comme
 
 1. Make sure that the very first cell of the notebook contains ONLY `# <title>` in it and has markdown type. This is important for link generation in the documentation.
 2. The `path` option must be a local path and cannot contain subfolders.
-3. The `description` option is used to pre-generate the `Examples` page overview in the documentation.
+3. The text in the `description` option will be used on the `Examples` page in the documentation.
 4. Set `hidden = true` option to hide a certain example in the documentation (the example will be executed to ensure it runs without errors).
 5. Please do no use `Overview` as a name for the new example, the `Overview` is reserved.
-6. Use the following template for formulas, not that `$$` and both `\begin` and `\end` commands are on the same line (check other examples if you are not sure). This is important, because otherwise formulas may not render correctly. Inline formulas may use `$...$` template.
+6. Use the following template for formulas, note that `$$` and both `\begin` and `\end` commands are on the same line (check other examples if you are not sure). This is important, because otherwise formulas may not render correctly. Inline formulas may use `$...$` template.
 ```
 $$\begin{aligned}
       <latex formulas here>
