### 2023-03-27 fb6ca47
Update new-example.md
