+++ docs/src/contributing/new-example.md
@@ -2,24 +2,24 @@
 
 We welcome all possible contributors. This page details some of the guidelines that should be followed when adding a new example (in the `examples/` folder) to this package.
 
-In order to add a new example simply create a new Jupyter notebook with your experiments in the `examples/` folder. When creating a new example add a descriptive explanation of your experiments, model specification, inference constraints decisions and add appropriate results analysis. For other people, it would also be useful if you write descriptive comments along your code.
+In order to add a new example simply create a new Jupyter notebook with your experiments in the `examples/` folder. When creating a new example add a descriptive explanation of your experiments, model specification, inference constraints decisions and add appropriate results analysis. We expect examples to be readable for the general public and therefore highly value descriptive comments. If a submitted example only contains code, we will kindly request some changes to improve the readability.
 
-After that it is necessary to modify the `examples/.meta.jl` file. See the comments in this file for more information.
+After preparing the new example it is necessary to modify the `examples/.meta.jl` file. See the comments in this file for more information.
 
-1. Make sure that the very first cell of the notebook contains ONLY `# <title>` in it and has markdown type. This is important for link generation in the documentation.
-2. The `path` option must be a local path and cannot contain subfolders.
+1. Make sure that the very first cell of the notebook contains ONLY `# <title>` in it and has the markdown cell type. This is important for generating links in our documentation.
+2. The `path` option must be set to a local path and cannot contain subfolders.
 3. The text in the `description` option will be used on the `Examples` page in the documentation.
 4. Set `hidden = true` option to hide a certain example in the documentation (the example will be executed to ensure it runs without errors).
-5. Please do no use `Overview` as a name for the new example, the `Overview` is reserved.
-6. Use the following template for formulas, note that `$$` and both `\begin` and `\end` commands are on the same line (check other examples if you are not sure). This is important, because otherwise formulas may not render correctly. Inline formulas may use `$...$` template.
+5. Please do no use `Overview` as a name for the new example, the title `Overview` is reserved.
+6. Use the following template for equations, note that `$$` and both `\begin` and `\end` commands are on the same line (check other examples if you are not sure). This is important, because otherwise formulas may not render correctly. Inline equations may use `$...$` template.
 ```
 $$\begin{aligned}
-      <latex formulas here>
+      <latex equations here>
 \end{aligned}$$
 ``` 
-7. Notebooks and plain Julia have different scoping rules for global variables, if it happens so that examples generation fails due to `UndefVarError` and scoping issues use `let ... end` blocks to enforce local scoping (see `Gaussian Mixtures Multivariate.ipynb` as an example)
+7. Notebooks and plain Julia have different scoping rules for global variables. It may happen that the generation of your example fails due to an `UndefVarError` or other scoping issues. In these cases we recommend using `let ... end` blocks to enforce local scoping (see `Gaussian Mixtures Multivariate.ipynb` as an example)
 
-8. All examples must use and activate local `Project.toml` in the second cell (see `1.`), if you need some package add it to the `(examples)` project
+8. All examples must use and activate the local environment specified by `Project.toml` in the second cell (see `1.`). Please have a look at the existing notebooks for an example on how to activate this local environment. If you need additional packages, you can add then to the `(examples)` project.
 
 !!! note
     Use `make examples` to run all examples or `make examples specific=MyNewCoolNotebook` to run any notebook that includes `MyNewCoolNotebook` in its file name.
