### 2023-03-28 #5161 @jmschrei
Sorry for making you change things again but the author list on the paper must match the author list on the Zenodo. Would you mind removing the names of those who are not included in the paper? Then, I think we're good to go.
