### 2023-03-29 #5161 @bvdmitri
Yeah, good point. The list was auto-generated from the list of contributors on GitHub. I have changed the list of authors manually on Zenodo. https://doi.org/10.5281/zenodo.7774921
