+++ Project.toml
@@ -2,7 +2,7 @@ name = "RxInfer"
 uuid = "86711068-29c9-4ff7-b620-ae75d7495b3d"
 authors = ["Bagaev Dmitry <d.v.bagaev@tue.nl> and contributors"]
 # Do not forget to update the version in the `codemeta.json` too
-version = "2.10.1"
+version = "2.10.2"
 
 [deps]
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
