### 2023-03-31 #5161 @bvdmitri
Sure. I'm working on it. The DOI itself has been fixed, but we have `PyPlot` related issues on Github Actions. This is release blocking (we cannot build the examples currently). I'll return to you once it is fixed.
