### 2023-03-31 8f8787f
examples: remove pyplot dependency
