### 2023-03-31 467a413
docs: update examples guide
