### 2023-03-31 #5161 @bvdmitri
@jmschrei Hi @jmschrei, just wanted to let you know that we have addressed the CI-related issues, and the latest release, version 2.10.2, is now available for our users. The record on [Zenodo](https://doi.org/10.5281/zenodo.7774921) has been updated to `2.10.2` version as well. This version includes the fix for the DOI in the paper.

I have a question though: in this issue the branch is set to `dev-joss-paper`. The paper itself has been merged to the `main` branch (for the Zenodo entry). Consequently, the fix for missing DOI is present on the `main` branch as well. Would it make sense to update the branch in this issue as well?
