### 2023-04-11 4458a26
fix windows commands. do not fail when examples are missing
