+++ docs/make.jl
@@ -5,15 +5,41 @@ using Documenter
 ## https://gr-framework.org/workstations.html#no-output
 ENV["GKSwstype"] = "100"
 
-DocMeta.setdocmeta!(RxInfer, :DocTestSetup, :(using RxInfer); recursive = true)
+DocMeta.setdocmeta!(RxInfer, :DocTestSetup, :(using RxInfer); recursive=true)
 
 # This must be auto-generated with `make examples`
 ExamplesPath = joinpath(@__DIR__, "src", "examples")
+ExamplesOverviewPath = joinpath(ExamplesPath, "overview.md")
 ExamplesMeta = include(joinpath(@__DIR__, "..", "examples", ".meta.jl"))
 Examples = map(filter(example -> !example.hidden, ExamplesMeta)) do examplemeta
     path = examplemeta.path
     title = examplemeta.title
     return title => joinpath("examples", replace(path, ".ipynb" => ".md"))
+end
+
+# Check if some examples are missing from the build
+ExistingExamples = filter(Examples) do (title, path)
+    exists = isfile(path)
+    if !exists
+        @warn "Example at path $(path) does not exist. Skipping."
+    end
+    return exists
+end
+
+if length(Examples) !== length(ExistingExamples)
+    @warn "Some examples were not found. Use the `make examples` command to generate all examples."
+end
+
+if !isdir(ExamplesPath)
+    mkpath(ExamplesPath)
+end
+
+# Check if the `overview.md` file exists
+if !isfile(ExamplesOverviewPath)
+    @warn "`$(ExamplesOverviewPath)` does not exist. Generating an empty overview. Use the `make examples` command to generate the overview and all examples."
+    open(ExamplesOverviewPath, "w") do overview
+        write(overview, "The overview is missing. Use the `make examples` command to generate the overview and all examples.")
+    end
 end
 
 # WIP: Keep it as a nice starting approach for adding a header, currently we are using `assets/header.js`
@@ -31,31 +57,31 @@ end
 # end
 
 makedocs(;
-    draft = false,
-    strict = [ :doctest, :eval_block, :example_block, :meta_block, :parse_error, :setup_block ],
-    modules = [ RxInfer ],
-    authors = "Bagaev Dmitry <d.v.bagaev@tue.nl> and contributors",
-    repo = "https://github.com/biaslab/RxInfer.jl/blob/{commit}{path}#{line}",
-    sitename = "RxInfer.jl",
-    format = Documenter.HTML(;
-        prettyurls = get(ENV, "CI", "false") == "true",
-        canonical = "https://biaslab.github.io/RxInfer.jl",
-        edit_link = "main",
-        assets = String[ "assets/theme.css", "assets/header.css", "assets/header.js" ],
+    draft=false,
+    strict=[:doctest, :eval_block, :example_block, :meta_block, :parse_error, :setup_block],
+    modules=[RxInfer],
+    authors="Bagaev Dmitry <d.v.bagaev@tue.nl> and contributors",
+    repo="https://github.com/biaslab/RxInfer.jl/blob/{commit}{path}#{line}",
+    sitename="RxInfer.jl",
+    format=Documenter.HTML(;
+        prettyurls=get(ENV, "CI", "false") == "true",
+        canonical="https://biaslab.github.io/RxInfer.jl",
+        edit_link="main",
+        assets=String["assets/theme.css", "assets/header.css", "assets/header.js"]
     ),
-    pages = [
-        "Home"     => "index.md",
-        "User guide"  => [
+    pages=[
+        "Home" => "index.md",
+        "User guide" => [
             # "Background: variational inference" => "manuals/background.md",
-            "Getting started"           => "manuals/getting-started.md",
-            "Model specification"       => "manuals/model-specification.md",
+            "Getting started" => "manuals/getting-started.md",
[…]
