### 2023-04-11 03abf7d
add debugging section to the documentation
