+++ docs/make.jl
@@ -51,6 +51,7 @@ makedocs(;
             "Model specification"       => "manuals/model-specification.md",
             "Constraints specification" => "manuals/constraints-specification.md",
             "Meta specification"        => "manuals/meta-specification.md",
+            "Debugging"                 => "manuals/debugging.md",
             "Inference specification"   => [ 
                 "Overview" => "manuals/inference/overview.md",
                 "Static dataset" => "manuals/inference/inference.md",
