+++ Makefile
@@ -5,9 +5,22 @@ SHELL = /bin/bash
 # - `RM` command: use for removing files
 ifeq ($(OS), Windows_NT) 
     RM = del /Q /F
+    PATH_SEP = \\
 else
     RM = rm -rf
+    PATH_SEP = /
 endif
+
+# Includes `examples/Manifest.toml`
+EXAMPLES_MANIFEST_FILE = examples$(PATH_SEP)Manifest.toml 
+# Includes `examples/Manifest.toml`, `docs/src/examples`, `docs/src/assets/examples`
+EXAMPLES_FILES = $(EXAMPLES_MANIFEST_FILE) docs$(PATH_SEP)src$(PATH_SEP)examples docs$(PATH_SEP)src$(PATH_SEP)assets$(PATH_SEP)examples
+# Includes `docs/build`
+DOCS_BUILD_FILES = docs$(PATH_SEP)build
+# Includes `_output`, `test/_output`
+TEST_OUTPUT_FILES = _output test$(PATH_SEP)_output
+# Includes all the above
+ALL_TMP_FILES = $(EXAMPLES_FILES) $(DOCS_BUILD_FILES) $(TEST_OUTPUT_FILES)
 
 .PHONY: lint format
 
@@ -23,11 +36,11 @@ format: scripts_init ## Code formating run
 .PHONY: examples
 
 examples_init:
-	$(RM) examples/Manifest.toml
+	$(RM) $(EXAMPLES_MANIFEST_FILE)
 	julia --startup-file=no --project=examples/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate(); Pkg.precompile();'
 
 dev_examples_init:
-	$(RM) examples/Manifest.toml
+	$(RM) $(EXAMPLES_MANIFEST_FILE)
 	julia --startup-file=no --project=examples/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.develop(PackageSpec(path=joinpath(Pkg.devdir(), "ReactiveMP.jl"))); Pkg.develop(PackageSpec(path=joinpath(Pkg.devdir(), "GraphPPL.jl"))); Pkg.develop(PackageSpec(path=joinpath(Pkg.devdir(), "Rocket.jl"))); Pkg.update(); Pkg.precompile();'
 
 examples: scripts_init examples_init ## Precompile examples and put them in the `docs/src/examples` folder (use specific="<pattern>" to compile a specific example)
@@ -39,11 +52,11 @@ devexamples: scripts_init dev_examples_init ## Same as `make examples` but uses 
 .PHONY: docs
 
 doc_init:
-	$(RM) docs/Manifest.toml
+	$(RM) $(EXAMPLES_MANIFEST_FILE)
 	julia --project=docs -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate(); Pkg.precompile();'
 
 dev_doc_init:
-	$(RM) docs/Manifest.toml
+	$(RM) $(EXAMPLES_MANIFEST_FILE)
 	julia --project=docs -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.develop(PackageSpec(path=joinpath(Pkg.devdir(), "ReactiveMP.jl"))); Pkg.develop(PackageSpec(path=joinpath(Pkg.devdir(), "GraphPPL.jl"))); Pkg.develop(PackageSpec(path=joinpath(Pkg.devdir(), "Rocket.jl"))); Pkg.update(); Pkg.precompile();'
 
 docs: doc_init ## Generate documentation
[…]
