### 2023-04-20 biaslab/RxInfer.jl#102 @bvdmitri
Add CITATION.cff file
https://github.com/openjournals/joss-reviews/issues/5161#issuecomment-1516475502
