### 2023-04-20 biaslab/RxInfer.jl#103 @bvdmitri
Add JOSS paper label
https://github.com/openjournals/joss-reviews/issues/5161#issuecomment-1516616453
